CI and Unit fixes
