+++ .circleci/config.yml
@@ -1,0 +1,23 @@
+version: 2
+jobs:
+  build:
+    docker:
+       - image: circleci/ruby:2.5.1-browsers
+
+    working_directory: ~/repo
+
+    steps:
+      - checkout
+      - run:
+          name: Install Danger
+          command: gem install danger
+      - run:
+          name: Run Danger
+          command: danger
+
+workflows:
+  version: 2
+  danger-flow:
+    jobs:
+      - build:
+          context: Danger
