+++ .foodcritic
@@ -1,3 +1,3 @@
 ~solo
 ~FC015
-~FC023
+~FC121
