--- .rspec
@@ -1,2 +1,0 @@
---color
---format progress
