+++ .rubocop.yml
@@ -1,11 +1,2 @@
-AllCops:
-  DisplayCopNames: true
-  Exclude:
-    - vendor/**/*
-
-# Cookstyle (rubocop) todo:
 BlockNesting:
-  Max: 4 # TODO: reduce to 3
-
-Style/EachWithObject:
-  Enabled: false
+  Max: 4
