+++ .travis.yml
@@ -1,15 +1,13 @@
 sudo: required
 dist: trusty
+cache: bundler
 
 addons:
   apt:
     sources:
-      - chef-current-trusty
+      - chef-stable-trusty
     packages:
       - chefdk
-
-# Don't `bundle install`
-install: echo "skip bundle install"
 
 branches:
   only:
@@ -18,38 +16,29 @@ branches:
 services: docker
 
 env:
-  matrix:
-  - INSTANCE=default-centos-6
-  - INSTANCE=default-centos-7
-  - INSTANCE=default-debian-7
-  - INSTANCE=default-debian-8
-  - INSTANCE=default-fedora-latest
-  - INSTANCE=default-ubuntu-1404
-  - INSTANCE=default-ubuntu-1604
-  - INSTANCE=mms-agent-centos-6
-  - INSTANCE=mms-agent-centos-7
-  - INSTANCE=mms-agent-ubuntu-1404
-  - INSTANCE=mms-agent-ubuntu-1604
+- INSTANCE=add-to-list-centos-7
+- INSTANCE=add-to-list-ubuntu-1604
+- INSTANCE=append-if-no-line-centos-7
+- INSTANCE=append-if-no-line-ubuntu-1604
+- INSTANCE=delete-from-list-centos-7
+- INSTANCE=delete-from-list-ubuntu-1604
+- INSTANCE=delete-lines-centos-7
+- INSTANCE=delete-lines-ubuntu-1604
+- INSTANCE=replace-or-add-centos-7
+- INSTANCE=replace-or-add-ubuntu-1604
 
-# Ensure we make ChefDK's Ruby the default
 before_script:
-  - eval "$(/opt/chefdk/bin/chef shell-init bash)"
-  - sudo iptables -L DOCKER || ( echo "DOCKER iptables chain missing" ; sudo iptables -N DOCKER )
-  - eval "$(/opt/chefdk/bin/chef shell-init bash)"
-  - /opt/chefdk/embedded/bin/chef --version
-  - /opt/chefdk/embedded/bin/cookstyle --version
-  - /opt/chefdk/embedded/bin/foodcritic --version
+- sudo iptables -L DOCKER || ( echo "DOCKER iptables chain missing" ; sudo iptables -N DOCKER )
+- eval "$(/opt/chefdk/bin/chef shell-init bash)"
+- chef --version
+- cookstyle --version
+- foodcritic --version
+- chef gem install danger
 
-script:
-  - KITCHEN_LOCAL_YAML=.kitchen.dokken.yml /opt/chefdk/embedded/bin/kitchen verify ${INSTANCE}
-
-after_failure:
-  - cat .kitchen/logs/*.log
+script: KITCHEN_LOCAL_YAML=.kitchen.dokken.yml kitchen verify ${INSTANCE}
 
 matrix:
-  allow_failures:
-    - env: INSTANCE=mms-agent-centos-6
   include:
     - script:
-      - /opt/chefdk/bin/chef exec rake
+      - chef exec delivery local all
       env: UNIT_AND_LINT=1
