--- NOTICE
@@ -1,9 +1,0 @@
-======================
-Chef Cookbooks Notices
-======================
-
-Developed at edelight GmbH (http://www.edelight-group.com/).
-
-Contributors:
-
- * Markus Korn <markus.korn@edelight.de>
